requirements.txt
Django==6.0
djangorestframework==3.14.0
drf-yasg==1.21.6
python-decouple==3.8
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
openai==1.3.0
